backend/requirements.txt
bcrypt==4.1.2
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.31.0
SQLAlchemy==2.0.23
